--- src/system/CollapsibleNotice/index.ts
@@ -1,6 +1,0 @@
-/**
- * Internal dependencies
- */
-import { CollapsibleNotice } from './CollapsibleNotice';
-
-export { CollapsibleNotice };
